+++ client/post-editor/test/post-editor.jsx
@@ -53,7 +53,7 @@ describe( 'PostEditor', function() {
 		mockery.registerMock( 'post-editor/editor-preview', MOCK_COMPONENT );
 		mockery.registerMock( 'post-editor/invalid-url-dialog', MOCK_COMPONENT );
 		mockery.registerMock( 'post-editor/restore-post-dialog', MOCK_COMPONENT );
-		mockery.registerMock( 'post-editor/editor-sidebar/header', MOCK_COMPONENT );
+		mockery.registerMock( 'post-editor/editor-sidebar', MOCK_COMPONENT );
 		mockery.registerMock( './editor-preview', MOCK_COMPONENT );
 		mockery.registerMock( 'my-sites/drafts/draft-list', MOCK_COMPONENT );
 		mockery.registerMock( 'lib/layout-focus', { set() {} } );
